Make YTtoLBRY use Preact
Not particularly valuable here, but more consistent if we're to accept
popup as using preact
